fix(http2): `max_header_list_size(num)` defaults to 16kb
The HTTP/2 does not define a default. If not defined, hyper still set a
high limit of 16mb. However, that seems very high, and most people
likely do not think to set it the property.

Since hyper tries to protect users, it will now use a default of 16kb.

The defaults in hyper are not part of the public API stability promise.
Users are encouraged to set options themselves.